support for sample type datasets
