+++ api/src/org/labkey/api/assay/AbstractAssayProvider.java
@@ -35,14 +35,12 @@ import org.labkey.api.data.ContainerManager;
 import org.labkey.api.data.DataRegion;
 import org.labkey.api.data.DetailsColumn;
 import org.labkey.api.data.DisplayColumn;
-import org.labkey.api.data.Filter;
 import org.labkey.api.data.RuntimeSQLException;
 import org.labkey.api.data.SQLFragment;
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.data.SqlSelector;
 import org.labkey.api.data.Table;
 import org.labkey.api.data.TableInfo;
-import org.labkey.api.data.TableSelector;
 import org.labkey.api.defaults.DefaultValueService;
 import org.labkey.api.exp.DomainNotFoundException;
 import org.labkey.api.exp.ExperimentException;
@@ -92,12 +90,10 @@ import org.labkey.api.security.permissions.AdminPermission;
 import org.labkey.api.security.permissions.DeletePermission;
 import org.labkey.api.settings.AppProps;
 import org.labkey.api.study.Dataset;
-import org.labkey.api.study.Study;
-import org.labkey.api.study.StudyService;
 import org.labkey.api.study.TimepointType;
-import org.labkey.api.study.publish.PublishKey;
-import org.labkey.api.study.assay.AssayPublishService;
 import org.labkey.api.study.assay.ParticipantVisitResolverType;
+import org.labkey.api.study.publish.PublishKey;
+import org.labkey.api.study.publish.StudyPublishService;
 import org.labkey.api.util.FileUtil;
 import org.labkey.api.util.HelpTopic;
 import org.labkey.api.util.PageFlowUtil;
@@ -244,7 +240,7 @@ public abstract class AbstractAssayProvider implements AssayProvider
                         targetStudyContainer = publishKey.getTargetStudy();
                     assert targetStudyContainer != null;
 
-                    TimepointType studyType = AssayPublishService.get().getTimepointType(targetStudyContainer);
+                    TimepointType studyType = StudyPublishService.get().getTimepointType(targetStudyContainer);
 
                     Map<String, Object> dataMap = new HashMap<>();
 
@@ -257,22 +253,22 @@ public abstract class AbstractAssayProvider implements AssayProvider
                         sourceContainer = ExperimentService.get().getExpRun(runLSID).getContainer();
                     }
 
-                    dataMap.put(AssayPublishService.PARTICIPANTID_PROPERTY_NAME, publishKey.getParticipantId());
+                    dataMap.put(StudyPublishService.PARTICIPANTID_PROPERTY_NAME, publishKey.getParticipantId());
 
                     if (!studyType.isVisitBased())
                     {
-                        dataMap.put(AssayPublishService.DATE_PROPERTY_NAME, publishKey.getDate());
+                        dataMap.put(StudyPublishService.DATE_PROPERTY_NAME, publishKey.getDate());
                     }
                     else
                     {
                         // add the sequencenum only for visit-based studies, a date based sequencenum will get calculated
                         // for date-based studies in the ETL layer
-                        dataMap.put(AssayPublishService.SEQUENCENUM_PROPERTY_NAME, publishKey.getVisitId());
+                        dataMap.put(StudyPublishService.SEQUENCENUM_PROPERTY_NAME, publishKey.getVisitId());
                     }
 
-                    dataMap.put(AssayPublishService.SOURCE_LSID_PROPERTY_NAME, sourceLSID);
+                    dataMap.put(StudyPublishService.SOURCE_LSID_PROPERTY_NAME, sourceLSID);
                     dataMap.put(getTableMetadata(protocol).getDatasetRowIdPropertyName(), publishKey.getDataId());
-                    dataMap.put(AssayPublishService.TARGET_STUDY_PROPERTY_NAME, targetStudyContainer);
+                    dataMap.put(StudyPublishService.TARGET_STUDY_PROPERTY_NAME, targetStudyContainer);
 
                     // Remember which rows we're planning to copy, partitioned by the target study
                     Set<Integer> rowIds = rowIdsByTargetContainer.get(targetStudyContainer);
@@ -286,48 +282,20 @@ public abstract class AbstractAssayProvider implements AssayProvider
                     dataMaps.add(dataMap);
                 }
 
-                checkForAlreadyCopiedRows(user, protocol, errors, rowIdsByTargetContainer);
-
+                StudyPublishService.get().checkForAlreadyCopiedRows(user, Pair.of(Dataset.PublishSource.Assay, protocol.getRowId()), errors, rowIdsByTargetContainer);
                 if (!errors.isEmpty())
                 {
                     return null;
                 }
                 
-                return AssayPublishService.get().publishAssayData(user, sourceContainer, study, protocol.getName(), protocol,
+                return StudyPublishService.get().publishData(user, sourceContainer, study, protocol.getName(),
+                        Pair.of(Dataset.PublishSource.Assay, protocol.getRowId()),
                         dataMaps, getTableMetadata(protocol).getDatasetRowIdPropertyName(), errors);
             }
         }
         catch (SQLException e)
         {
             throw new RuntimeSQLException(e);
-        }
-    }
-
-    private void checkForAlreadyCopiedRows(User user, ExpProtocol protocol, List<String> errors, Map<Container, Set<Integer>> rowIdsByTargetContainer)
-    {
-        for (Map.Entry<Container, Set<Integer>> entry : rowIdsByTargetContainer.entrySet())
-        {
-            Study targetStudy = StudyService.get().getStudy(entry.getKey());
-
-            // Look for an existing dataset backed by this assay definition
-            for (Dataset dataset : targetStudy.getDatasets())
-            {
-                if (protocol.equals(dataset.getAssayProtocol()) && dataset.getKeyPropertyName() != null)
-                {
-                    // Check to see if it already has the data rows that are being copied
-                    TableInfo tableInfo = dataset.getTableInfo(user, false);
-                    Filter datasetFilter = new SimpleFilter(new SimpleFilter.InClause(FieldKey.fromParts(dataset.getKeyPropertyName()), entry.getValue()));
-                    long existingRowCount = new TableSelector(tableInfo, datasetFilter, null).getRowCount();
-                    if (existingRowCount > 0)
-                    {
-                        // If so, don't let the user copy them again, even if they have different participant/visit/date info
-                        String errorMessage = existingRowCount == 1 ? "One of the selected rows has" : (existingRowCount + " of the selected rows have");
-                        errorMessage += " already been copied to the study \"" + targetStudy.getLabel() + "\" in " + entry.getKey().getPath();
-                        errorMessage += " (RowIds: " + entry.getValue() + ")";
-                        errors.add(errorMessage);
-                    }
-                }
-            }
         }
     }
 
@@ -1547,7 +1515,7 @@ public abstract class AbstractAssayProvider implements AssayProvider
             result.add(new NavTree("view upload jobs", PageFlowUtil.addLastFilterParameter(url, AssayProtocolSchema.getLastFilterScope(protocol))));
         }
 
-        AssayPublishService svc = AssayPublishService.get();
+        StudyPublishService svc = StudyPublishService.get();
 
         if (AuditLogService.get().isViewable() && null != svc)
             result.add(new NavTree("view copy-to-study history", svc.getPublishHistory(viewContext.getContainer(), protocol, containerFilter)));
