+++ api/src/org/labkey/api/qc/export/QCStateImportExportHelper.java
@@ -54,7 +54,10 @@ public interface QCStateImportExportHelper
     QCState insertQCState(User user, QCState state);
     QCState updateQCState(User user, QCState state);
 
-    void setDefaultAssayQCState(Container container, User user, Integer stateId);
+    /**
+     * The default QC state for data linked (published) to the study
+     */
+    void setDefaultPublishedDataQCState(Container container, User user, Integer stateId);
     void setDefaultPipelineQCState(Container container, User user, Integer stateId);
     void setDefaultDirectEntryQCState(Container container, User user, Integer stateId);
     void setShowPrivateDataByDefault(Container container, User user, boolean showPrivate);
