+++ study/resources/schemas/dbscripts/postgresql/study-21.002-21.003.sql
@@ -1,0 +1,4 @@
+ALTER TABLE study.Study RENAME COLUMN DefaultAssayQCState TO DefaultPublishDataQCState;
+
+ALTER TABLE study.Dataset RENAME COLUMN ProtocolId TO PublishSourceId;
+ALTER TABLE study.Dataset ADD COLUMN PublishSourceType VARCHAR(50) NOT NULL DEFAULT 'Assay';
