+++ study/resources/schemas/dbscripts/sqlserver/study-21.002-21.003.sql
@@ -1,0 +1,4 @@
+sp_rename 'study.Study.DefaultAssayQCState', 'DefaultPublishDataQCState', 'COLUMN';
+
+sp_rename 'study.Dataset.ProtocolId', 'PublishSourceId', 'COLUMN';
+ALTER TABLE study.Dataset ADD PublishSourceType NVARCHAR(50) NOT NULL DEFAULT 'Assay';
