+++ study/src/org/labkey/study/query/DatasetAutoJoinTable.java
@@ -162,7 +162,7 @@ public class DatasetAutoJoinTable extends VirtualTable
                 // A -> A
                 fk = createParticipantFK(dsd, cf);
         }
-        else if (_keyPropertyName == null || _source.isAssayData())
+        else if (_keyPropertyName == null || _source.isPublishedData())
         {
             if (dsd.isDemographicData())
                 // B -> A
@@ -176,7 +176,7 @@ public class DatasetAutoJoinTable extends VirtualTable
             if (dsd.isDemographicData())
                 // C -> A
                 fk = createParticipantFK(dsd, cf);
-            else if (dsd.getKeyPropertyName() == null || dsd.isAssayData())
+            else if (dsd.getKeyPropertyName() == null || dsd.isPublishedData())
                 // C -> B
                 fk = createParticipantSequenceNumFK(dsd, cf);
             else
@@ -210,7 +210,7 @@ public class DatasetAutoJoinTable extends VirtualTable
 
     private DatasetForeignKey createParticipantSequenceNumFK(DatasetDefinition dsd, ContainerFilter cf)
     {
-        assert !dsd.isDemographicData() && (dsd.getKeyPropertyName() == null || dsd.isAssayData());
+        assert !dsd.isDemographicData() && (dsd.getKeyPropertyName() == null || dsd.isPublishedData());
         assert !_source.isDemographicData();
 
         DatasetForeignKey fk = new DatasetForeignKey(dsd, cf);
